Estimate separate PCB effects for Puyallup and White
This commit uses the subyearling SAR of the model to calculate the size of juveniles in July, for use as the "base case" to convert the absolute survival rate of Duffy and Beauchamp to a relative change in survival due to change in size. Because the Puyallup and White models have different subyearling SARs, the growth-to-survival effects are different for each system.