Scratch code for exploring divergent integrals
Outliers in the population exposure distributions, particularly those with fat tails, tend to result in effect integrals that do not converge. There appears to be a fairly clear frontier where this occurs.

If extending this, consider calculating the kurtosis at each point in the grid to see if this determines convergence success or failure. Maybe there is a kurtosis-limiting prior that can be applied to the population fitting model?