Allow effect of remediating PCBs to be estimated (#14)
* Add option to "remove" PCBs

* Estimate separate PCB effects for Puyallup and White

This commit uses the subyearling SAR of the model to calculate the size of juveniles in July, for use as the "base case" to convert the absolute survival rate of Duffy and Beauchamp to a relative change in survival due to change in size. Because the Puyallup and White models have different subyearling SARs, the growth-to-survival effects are different for each system.

* Fix minor bugs in Puyallup and White population models

1. Prespawn migrants are tracked separately by age class so that age-specific fecundity can be applied. Originally, the Beverton-Holt production function was applied to each age class independently rather than collectively, reducing the impact of density-dependence. This commit applies the Beverton-Holt function to the sum of the age class abundances and assumes that the proportion in each age class is constant.

2. Previously, `spawners` was calculated from `migrantprespawn` when it should have been based on `holdingprespawn`.

* Scratch code for exploring divergent integrals

Outliers in the population exposure distributions, particularly those with fat tails, tend to result in effect integrals that do not converge. There appears to be a fairly clear frontier where this occurs.

If extending this, consider calculating the kurtosis at each point in the grid to see if this determines convergence success or failure. Maybe there is a kurtosis-limiting prior that can be applied to the population fitting model?